added

+++ server/controllers/postControllers.js
@@ -1,18 +1,28 @@
 const Post = require("../models/postModel")
-exports.createPost = async(req,res)=>{
+const cloudinary = require('../helpers/cloudinary');
+//add Post
+exports.createPost = async(req,res)=>{ 
     try {
-    //     const userId =req.userId
-    
-    //   const {title,desc,img}= req.body
-        const newPost = await Post.create(req.body)
+        // const newBody = JSON.parse(req.body.info)
+        // console.log(newBody)
+        // const imageInfo = await cloudinary.uploader.upload(req.file.path);
+       const userId = req.personId
+  
+       const {title,desc,img}= req.body
+        const newPost = await Post.create({
+            title,desc,img,
+            // image: { imageURL: imageInfo.url, public_id: imageInfo.public_id },
+            owner:userId})
+        console.log(req.body)
         res.json(newPost)
     } catch (error) {
+        console.log(error)
         res.status(500).json({msg:"something went wrong"})
     }
 }
 exports.getPosts = async(req,res)=>{
     try {
-        const posts = await Post.find().populate("owner").populate("likes")
+        const posts = await Post.find({}).populate("owner").populate("likes").populate("comments.commentOwner")
         res.json(posts)
     } catch (error) {
         res.status(500).json({msg:"something went wrong"})
@@ -34,14 +44,14 @@ exports.getPostById = async(req,res)=>{
         res.status(500).json({msg:"something went wrong"})
     }
 }
-exports.updatePost = async(req,res)=>{
-    try {
-        const updatedPost = await Post.findByIdAndUpdate(req.params.postId,req.body,{new:true})
-        res.json(updatedPost)
-    } catch (error) {
-        res.status(500).json({msg:"something went wrong"})
-    }
-}
+// exports.updatePost = async(req,res)=>{
+//     try {
+//         const updatedPost = await Post.findByIdAndUpdate(req.params.postId,req.body,{new:true})
+//         res.json(updatedPost)
+//     } catch (error) {
+//         res.status(500).json({msg:"something went wrong"})
+//     }
+// }
 exports.likePost = async(req,res)=>{
     try {
         const posts = await Post.findById(req.params.postId)
@@ -61,10 +71,48 @@ exports.likePost = async(req,res)=>{
         res.status(500).json({msg:"something went wrong"})
     }
 }
+//like post
+exports.likePosts = async(req,res)=>{
+    try {
+        const postId = req.params.postId
+        const posts = await Post.findById(postId)
+        console.log(req.params.postId,req.personId)
+        const chekLike = await posts.likes.find((el) => el == req.personId)
+       
+        if(chekLike) await  Post.findByIdAndUpdate(postId,{  $pull :{ likes :req.personId}});
+        else  await Post.findByIdAndUpdate(postId,{  $push :{ likes :req.personId}})
+        res.json({msg:"post is liked"})
+    } catch (error) {
+        console.log(error)
+        res.status(500).json({msg:"something went wrong"})
+    }
+}
+
+
 exports.deletePost = async(req,res)=>{
     try {
-        const deletedPost = await Post.findByIdAndDelete(req.params.postId)
-        res.json({msg:"post deleted with success",deletedPost})
+        
+        const post = await Post.findById(req.params.postId)
+        
+        console.log(String(post.owner._id))
+        console.log(req.personId)
+       if(String(post.owner._id) !== req.personId) return   res.status(401).json({msg:"not authorizied"})
+       await Post.findByIdAndDelete(req.params.postId)
+        res.json({msg:"post deleted with success"})
+    } catch (error) {
+        res.status(500).json({msg:"something went wrong"})
+    }
+}
+exports.updatePost = async(req,res)=>{
+    try {
+        
+        const post = await Post.findById(req.params.postId)
+        
+        console.log(String(post.owner._id))
+        console.log(req.personId)
+       if(String(post.owner._id) !== req.personId) return   res.status(401).json({msg:"not authorizied"})
+       await Post.findByIdAndUpdate(req.params.postId,{...req.body})
+        res.json({msg:"post updated with success"})
     } catch (error) {
         res.status(500).json({msg:"something went wrong"})
     }
